Audit.NET to net462;net472;netstandard2.0;netstandard2.1;net6.0;net7.0;net8.0
