+++ src/Audit.EntityFramework.Core/Properties/AssemblyInfo.cs
@@ -18,4 +18,5 @@ using System.Runtime.InteropServices;
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("f4b17a62-a499-4070-ac31-a812a9a048b0")]
 
-[assembly: InternalsVisibleTo("Audit.UnitTest,PublicKey=00240000048000009400000006020000002400005253413100040000010001007f14e426ecfadb7a221dd166152164af7395256b474a50d56dad8080bba2edbea4ca85d89401245686311e557c9853c47095c6602a32d0a207ba2a55d282aa69d4f54d58eaf1e876fa18ac58bcd93bbb2aea797d7efda08ca0bc910dca508acc6ab6dbafcf7fd5b63859582ba558bf14d09f0b13d7a18b26688c1b2b98fcddc9")]
+[assembly: InternalsVisibleTo("Audit.UnitTest,PublicKey=00240000048000009400000006020000002400005253413100040000010001007f14e426ecfadb7a221dd166152164af7395256b474a50d56dad8080bba2edbea4ca85d89401245686311e557c9853c47095c6602a32d0a207ba2a55d282aa69d4f54d58eaf1e876fa18ac58bcd93bbb2aea797d7efda08ca0bc910dca508acc6ab6dbafcf7fd5b63859582ba558bf14d09f0b13d7a18b26688c1b2b98fcddc9"),
+           InternalsVisibleTo("Audit.EntityFramework.Core.UnitTest,PublicKey=00240000048000009400000006020000002400005253413100040000010001007f14e426ecfadb7a221dd166152164af7395256b474a50d56dad8080bba2edbea4ca85d89401245686311e557c9853c47095c6602a32d0a207ba2a55d282aa69d4f54d58eaf1e876fa18ac58bcd93bbb2aea797d7efda08ca0bc910dca508acc6ab6dbafcf7fd5b63859582ba558bf14d09f0b13d7a18b26688c1b2b98fcddc9")]
