+++ src/Audit.Mvc/Audit.Mvc.csproj
@@ -44,7 +44,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNet.Mvc" Version="5.3.0" />
-		<PackageReference Include="System.Text.Json" Version="8.0.0" />
+		<PackageReference Include="System.Text.Json" Version="6.0.9" />
 		<Reference Include="System.Web" />
 		<Reference Include="System" />
 		<Reference Include="Microsoft.CSharp" />
