+++ src/Audit.NET/AuditActivityTrace.cs
@@ -1,4 +1,4 @@
-﻿#if NET5_0_OR_GREATER
+﻿#if NET6_0_OR_GREATER
 using System;
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
