+++ src/Audit.NET/IAuditScope.cs
@@ -4,10 +4,7 @@ using System.Threading.Tasks;
 
 namespace Audit.Core
 {
-    public interface IAuditScope : IDisposable
-#if NETSTANDARD2_0 || NETSTANDARD2_1 || NET461 || NET5_0_OR_GREATER
-        , IAsyncDisposable
-#endif
+    public interface IAuditScope : IDisposable, IAsyncDisposable
     {
         AuditDataProvider DataProvider { get; }
         AuditEvent Event { get; }
@@ -18,9 +15,6 @@ namespace Audit.Core
         void Comment(string text);
         void Comment(string format, params object[] args);
         void Discard();
-#if NET45 || NETSTANDARD1_3
-        ValueTask DisposeAsync();
-#endif
         void Save();
         Task SaveAsync(CancellationToken cancellationToken = default);
         void SetCustomField<TC>(string fieldName, TC value, bool serialize = false);
