+++ test/Audit.UnitTest/Audit.UnitTest.csproj
@@ -1,83 +1,48 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <VersionPrefix>3.0.0</VersionPrefix>
-    <TargetFrameworks>netcoreapp3.0;netcoreapp2.0;netcoreapp1.0;net461;net5.0;net6.0;net7.0;net8.0</TargetFrameworks>
-    <LangVersion>8.0</LangVersion>
-    <DefineConstants>$(DefineConstants);STRONG_NAME</DefineConstants>
-    <NoWarn>$(NoWarn);1591;NETSDK1138</NoWarn>
-    <AssemblyName>Audit.UnitTest</AssemblyName>
-    <AssemblyOriginatorKeyFile>../../src/StrongName/Audit.NET.UnitTests.snk</AssemblyOriginatorKeyFile>
-    <SignAssembly>true</SignAssembly>
-    <PublicSign Condition=" '$(OS)' != 'Windows_NT' ">true</PublicSign>
-    <PackageId>Audit.UnitTest</PackageId>
-    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-    <RuntimeFrameworkVersion Condition=" '$(TargetFramework)' == 'netcoreapp2.0' ">2.0.0</RuntimeFrameworkVersion>
-    <GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
-    <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
-    <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
-    <DebugType Condition=" '$(TargetFramework)' != 'netcoreapp2.0' ">Full</DebugType>
-  </PropertyGroup>
+	<PropertyGroup>
+		<VersionPrefix>3.0.0</VersionPrefix>
+		<TargetFrameworks>net462;net472;netcoreapp3.1;net6.0;net7.0;net8.0</TargetFrameworks>
+		<LangVersion>8.0</LangVersion>
+		<DefineConstants>$(DefineConstants);STRONG_NAME</DefineConstants>
+		<NoWarn>$(NoWarn);1591;NETSDK1138</NoWarn>
+		<AssemblyName>Audit.UnitTest</AssemblyName>
+		<AssemblyOriginatorKeyFile>../../src/StrongName/Audit.NET.UnitTests.snk</AssemblyOriginatorKeyFile>
+		<SignAssembly>true</SignAssembly>
+		<PublicSign Condition=" '$(OS)' != 'Windows_NT' ">true</PublicSign>
+		<PackageId>Audit.UnitTest</PackageId>
+		<GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
+		<GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
+		<GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
+		<GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
+		<DebugType>Full</DebugType>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Audit.NET.Kafka\Audit.NET.Kafka.csproj" />
-    <ProjectReference Include="..\..\src\Audit.NET.log4net\Audit.NET.log4net.csproj" />
-    <ProjectReference Include="..\..\src\Audit.NET.NLog\Audit.NET.NLog.csproj" />
-    <ProjectReference Include="..\..\src\Audit.NET.Serilog\Audit.NET.Serilog.csproj" />
-    <ProjectReference Include="..\..\src\Audit.NET\Audit.NET.csproj" />
-    <ProjectReference Include="..\..\src\Audit.NET.Udp\Audit.NET.Udp.csproj" />
-    <ProjectReference Include="..\..\src\Audit.EntityFramework\Audit.EntityFramework.csproj" />
-  </ItemGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' != 'netcoreapp1.0' ">
-    <ProjectReference Include="..\..\src\Audit.NET.ElasticSearch\Audit.NET.Elasticsearch.csproj" />
-  </ItemGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' != 'net5.0' AND '$(TargetFramework)' != 'net6.0' AND '$(TargetFramework)' != 'net7.0' AND '$(TargetFramework)' != 'net8.0' ">
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-  </ItemGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net5.0' ">
-	<PackageReference Include="System.Text.Json" Version="5.0.2" />
-  </ItemGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net5.0' OR '$(TargetFramework)' == 'net6.0' OR '$(TargetFramework)' == 'net7.0' OR '$(TargetFramework)' == 'net8.0' ">
-    <ProjectReference Include="..\..\src\Audit.NET.EventLog.Core\Audit.NET.EventLog.Core.csproj" />
-    <PackageReference Include="Microsoft.NETCore.Platforms" Version="5.0.2" />
-  </ItemGroup>
-  <PropertyGroup Condition=" '$(TargetFramework)' != 'net5.0' AND '$(TargetFramework)' != 'net6.0' AND '$(TargetFramework)' != 'net7.0' AND '$(TargetFramework)' != 'net8.0' ">
-    <DefineConstants>$(DefineConstants);NK_JSON</DefineConstants>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(TargetFramework)' == 'net5.0' OR '$(TargetFramework)' == 'net6.0' OR '$(TargetFramework)' == 'net7.0' OR '$(TargetFramework)' == 'net8.0' ">
-    <DefineConstants>$(DefineConstants);TEXT_JSON</DefineConstants>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NLog" Version="4.7.10" />
-    <PackageReference Include="log4net" Version="2.0.15" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
-    <PackageReference Include="NUnit" Version="3.7.1" />
-    <PackageReference Include="NUnit3TestAdapter" Version="3.8.0" />
-    <PackageReference Include="Moq" Version="4.7.99" />
-    <PackageReference Include="System.Net.Http" Version="4.3.4" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\..\src\Audit.NET.log4net\Audit.NET.log4net.csproj" />
+		<ProjectReference Include="..\..\src\Audit.NET.NLog\Audit.NET.NLog.csproj" />
+		<ProjectReference Include="..\..\src\Audit.NET\Audit.NET.csproj" />
+		<ProjectReference Include="..\..\src\Audit.NET.Udp\Audit.NET.Udp.csproj" />
+	</ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netcoreapp2.0' ">
-    <ProjectReference Include="..\..\src\Audit.NET.EventLog.Core\Audit.NET.EventLog.Core.csproj" />
-    <PackageReference Include="System.Runtime" Version="4.3.0" />
-    <PackageReference Include="Microsoft.NETCore.Platforms" Version="5.0.2" />
-  </ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net461' ">
-    <PackageReference Include="System.Runtime" Version="4.1.0" />
-    <PackageReference Include="Microsoft.AspNet.SignalR.Core" Version="2.1.2" />
-    <ProjectReference Include="..\..\src\Audit.SignalR\Audit.SignalR.csproj" />
-    <Reference Include="System" />
-    <Reference Include="Microsoft.CSharp" />
-  </ItemGroup>
+	<ItemGroup Condition=" '$(TargetFramework)' == 'net462' or '$(TargetFramework)' == 'net472' or '$(TargetFramework)' == 'netcoreapp3.1' ">
+		<PackageReference Include="System.Text.Json" Version="6.0.9" />
+	</ItemGroup>
 
-  <ItemGroup Condition="'$(TargetFramework)' != 'netcoreapp1.0'">
-    <PackageReference Include="FluentAssertions">
-      <Version>6.7.0</Version>
-    </PackageReference>
-    <PackageReference Include="Serilog.Sinks.InMemory.Assertions">
-      <Version>0.11.0</Version>
-    </PackageReference>
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
+		<PackageReference Include="NUnit" Version="3.7.1" />
+		<PackageReference Include="NUnit3TestAdapter" Version="3.8.0" />
+		<PackageReference Include="Moq" Version="4.7.99" />
+		<PackageReference Include="System.Net.Http" Version="4.3.4" />
+	</ItemGroup>
+
+	<ItemGroup Condition=" '$(TargetFramework)' == 'net462' ">
+		<PackageReference Include="System.Runtime" Version="4.1.0" />
+		<Reference Include="Microsoft.CSharp" />
+		<PackageReference Include="Microsoft.AspNet.SignalR.Core" Version="2.4.3" />
+		<ProjectReference Include="..\..\src\Audit.SignalR\Audit.SignalR.csproj" />
+	</ItemGroup>
 
 </Project>
