+++ test/Audit.UnitTest/SignalrTests.cs
@@ -1,17 +1,14 @@
-﻿#if NET461
+﻿#if NET462
 using NUnit.Framework;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Audit.Core;
-using Audit.Core.Extensions;
 using Audit.Core.Providers;
 using Audit.SignalR;
-using Castle.Components.DictionaryAdapter;
 using Microsoft.AspNet.SignalR;
 using Microsoft.AspNet.SignalR.Hubs;
 using Moq;
