+++ test/Audit.UnitTest/UnitTest.Async.cs
@@ -2,10 +2,7 @@
 using System;
 using Moq;
 using Audit.Core.Providers;
-using Audit.EntityFramework;
 using System.Collections.Generic;
-using Audit.Core.Extensions;
-using System.Diagnostics;
 using NUnit.Framework;
 using System.Threading.Tasks;
 using System.IO;
@@ -217,7 +214,7 @@ namespace Audit.UnitTest
                  .EventType("event type")
                  .ExtraFields(new { f = 1 })
                  .CreationPolicy(EventCreationPolicy.Manual)
-                 .AuditEvent(new AuditEventEntityFramework())
+                 .AuditEvent(new AuditEvent())
                  .DataProvider(new DynamicDataProvider())
                  .IsCreateAndSave(true)
                  .SkipExtraFrames(1)
@@ -228,7 +225,7 @@ namespace Audit.UnitTest
             Assert.AreEqual("event type", scope.Event.EventType);
             Assert.IsTrue(scope.Event.CustomFields.ContainsKey("f"));
             Assert.AreEqual(EventCreationPolicy.Manual, scope.EventCreationPolicy);
-            Assert.AreEqual(typeof(AuditEventEntityFramework), scope.Event.GetType());
+            Assert.AreEqual(typeof(AuditEvent), scope.Event.GetType());
             Assert.AreEqual(typeof(DynamicDataProvider), scope.DataProvider.GetType());
             Assert.AreEqual(SaveMode.InsertOnStart, scope.SaveMode);
             Assert.AreEqual("1", scope.Event.Target.Old.ToString());
@@ -243,7 +240,7 @@ namespace Audit.UnitTest
                  .EventType("event type")
                  .ExtraFields(new { f = 1 })
                  .CreationPolicy(EventCreationPolicy.Manual)
-                 .AuditEvent(new AuditEventEntityFramework())
+                 .AuditEvent(new AuditEvent())
                  .DataProvider(new DynamicDataProvider())
                  .IsCreateAndSave(true)
                  .SkipExtraFrames(1)
@@ -254,7 +251,7 @@ namespace Audit.UnitTest
             Assert.AreEqual("event type", scope.Event.EventType);
             Assert.IsTrue(scope.Event.CustomFields.ContainsKey("f"));
             Assert.AreEqual(EventCreationPolicy.Manual, scope.EventCreationPolicy);
-            Assert.AreEqual(typeof(AuditEventEntityFramework), scope.Event.GetType());
+            Assert.AreEqual(typeof(AuditEvent), scope.Event.GetType());
             Assert.AreEqual(typeof(DynamicDataProvider), scope.DataProvider.GetType());
             Assert.AreEqual(SaveMode.InsertOnStart, scope.SaveMode);
             Assert.AreEqual("1", scope.Event.Target.Old.ToString());
@@ -711,7 +708,6 @@ namespace Audit.UnitTest
             provider.Verify(p => p.InsertEvent(It.IsAny<AuditEvent>()), Times.Exactly(1));
         }
 
-#if NETCOREAPP2_0 || NETCOREAPP3_0 || NET5_0_OR_GREATER
         [Test]
         public async Task Test_Dispose_Async()
         {
@@ -723,7 +719,6 @@ namespace Audit.UnitTest
 
             provider.Verify(p => p.InsertEventAsync(It.IsAny<AuditEvent>(), It.IsAny<CancellationToken>()), Times.Exactly(1));
         }
-#endif
 
         [Test]
         public async Task TestDiscard_Async()
@@ -848,20 +843,6 @@ namespace Audit.UnitTest
             Assert.NotNull(scope2.EventId);
             Assert.AreNotEqual(scope1.EventId, scope2.EventId);
             provider.Verify(p => p.InsertEventAsync(It.IsAny<AuditEvent>(), It.IsAny<CancellationToken>()), Times.Exactly(2));
-        }
-
-        [AuditDbContext(AuditEventType = "FromAttr")]
-        public class MyContext : AuditDbContext
-        {
-            public override string AuditEventType { get { return base.AuditEventType; } }
-            public override bool IncludeEntityObjects { get { return base.IncludeEntityObjects; } }
-            public override AuditOptionMode Mode { get { return base.Mode; } }
-        }
-        public class AnotherContext : AuditDbContext
-        {
-            public override string AuditEventType { get { return base.AuditEventType; } }
-            public override bool IncludeEntityObjects { get { return base.IncludeEntityObjects; } }
-            public override AuditOptionMode Mode { get { return base.Mode; } }
         }
     }
 }
